DB-4070:
- Add dts vite plugin to wordpress-kit
- Remove rollup config
- Add build config to vite.config.js
- Restructure directories to better align with nextjs-kit
- Add some missing TSDoc comments
