+++ packages/wordpress-kit/index.ts
@@ -1,9 +1,5 @@
-import GraphqlClientFactory from './src/lib/GraphqlClientFactory';
-import tailwindcssPlugin from './src/lib/TailwindcssPlugin';
 import { gql } from 'graphql-request';
-import setEdgeHeader from './src/lib/setEdgeHeader';
 
-import type { TailwindcssConfig } from './src/types/TailwindcssPlugin';
-
-export { GraphqlClientFactory, gql, tailwindcssPlugin, setEdgeHeader };
-export type { TailwindcssConfig };
+export { gql };
+export * from './src/lib';
+export * from './src/types';
