+++ packages/wordpress-kit/src/lib/GraphqlClientFactory.ts
@@ -1,8 +1,19 @@
 import { GraphQLClient } from 'graphql-request';
 import type { PatchedRequestInit } from 'graphql-request/dist/types';
 
+/**
+ * Creates instances of `graphql-request` GraphQLClient with the given options
+ * @params endpoint - A WordPress GraphQL endpoint
+ * @params options - A RequestInit object. {@link PatchedRequestInit}
+ */
 class GraphQLClientFactory {
+	/**
+	 * A WordPress GraphQL endpoint
+	 */
 	endpoint: string;
+	/**
+	 * {@link PatchedRequestInit}
+	 */
 	options: PatchedRequestInit;
 
 	constructor(endpoint: string, options: PatchedRequestInit = {}) {
@@ -10,6 +21,9 @@ class GraphQLClientFactory {
 		this.options = options;
 	}
 
+	/**
+	 * Creates an instance of `graphql-request` GraphQLClient based on the endpoint and options
+	 */
 	create(): GraphQLClient {
 		return new GraphQLClient(this.endpoint, this.options);
 	}
