+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/utilities/Colors.ts
@@ -1,8 +1,4 @@
-import type {
-	GradientColors,
-	ThemeType,
-	Color,
-} from '../../../types/TailwindcssPlugin';
+import type { GradientColors, ThemeType, Color } from '../../../types';
 import { colorList, gradientList } from '../Constants';
 
 /**
