+++ packages/wordpress-kit/vite.config.js
@@ -1,11 +1,30 @@
-const path = require('path');
+import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
 
-module.exports = {
-  build: {
-    lib: {
-      entry: path.resolve(__dirname, 'index.ts'),
-      name: 'WordPressKit',
-      fileName: format => `wordpress-kit.${format}.js`,
-    },
-  },
+const globals = {
+	'graphql-request': 'graphqlRequest',
+	'tailwindcss/plugin': 'tailwindcssPlugin',
 };
+
+const external = ['graphql-request', 'tailwindcss/plugin'];
+
+/** @type {import('vite').defineConfig} */
+export default defineConfig(() => {
+	return {
+		plugins: [dts({ insertTypesEntry: true })],
+		build: {
+			lib: {
+				entry: './index.ts',
+				name: 'wordpress-kit',
+				formats: ['umd', 'es'],
+				fileName: (format) => `wordpress-kit.${format}.js`,
+			},
+			rollupOptions: {
+				external,
+				output: {
+					globals,
+				},
+			},
+		},
+	};
+});
